move threat index patterns validator to a separate folder
