+++ x-pack/plugins/security_solution/public/detection_engine/rule_creation/components/threat_match_index_edit/validators/translations.tsx
@@ -1,0 +1,15 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { i18n } from '@kbn/i18n';
+
+export const THREAT_MATCH_INDEX_FIELD_VALIDATION_REQUIRED_ERROR = i18n.translate(
+  'xpack.securitySolution.ruleManagement.ruleFields.threatMatchIndex.validation.requiredError',
+  {
+    defaultMessage: 'A minimum of one index pattern is required.',
+  }
+);
